moving prometheus_containers to modules
