+++ services/resource-usage-tracker/tests/unit/with_dbs/test_collect_container_resource_usage_task.py
@@ -7,7 +7,7 @@ import sqlalchemy as sa
 from fastapi import FastAPI
 from pytest_mock import MockerFixture
 from simcore_postgres_database.models.resource_tracker import resource_tracker_container
-from simcore_service_resource_usage_tracker.prometheus_containers.core import (
+from simcore_service_resource_usage_tracker.modules.prometheus_containers.core import (
     collect_container_resource_usage_task,
 )
 
@@ -34,7 +34,7 @@ def mocked_prometheus_client_custom_query(
         data = json.load(file)
 
     mocked_get_prometheus_api_client = mocker.patch(
-        "simcore_service_resource_usage_tracker.prometheus_containers.core._prometheus_sync_client_custom_query",
+        "simcore_service_resource_usage_tracker.modules.prometheus_containers.core._prometheus_sync_client_custom_query",
         autospec=True,
         return_value=data,
     )
